@@ -26,8 +26,8 @@
     "express": "^4.17.1",
     "express-fileupload": "^1.4.0",
     "imagemagick": "^0.1.3",
-    "ipfs-http-client": "56.0.2",
     "jsonwebtoken": "^8.5.1",
+    "kubo-rpc-client": "^3.0.4",
     "morgan": "^1.10.0",
     "node-latex": "^3.1.0",
     "reflect-metadata": "^0.1.13"
@@ -38,6 +38,7 @@
     "@types/body-parser": "^1.19.1",
     "@types/cors": "^2.8.12",
     "@types/express": "^4.17.13",
+    "@types/imagemagick": "^0.0.35",
     "@types/jsonwebtoken": "^8.5.4",
     "@types/morgan": "^1.9.3",
     "@types/node": "^16.4.13",
@@ -52,14 +53,13 @@
     "eslint-plugin-import": "^2.24.0",
     "eslint-plugin-no-array-reduce": "^1.0.22",
     "eslint-plugin-prettier": "^3.4.0",
-    "ipfs-core": "^0.15.4",
     "lint-staged": "^11.1.2",
     "prettier": "^2.3.2",
     "pretty-quick": "^3.1.1",
     "rimraf": "^3.0.2",
     "ts-node": "10.2.0",
     "ts-node-dev": "^1.1.8",
-    "typescript": "4.3.5"
+    "typescript": "^5.0.0"
   },
   "lint-staged": {
     "*": [
