@@ -59,6 +59,7 @@ import {
   createNotEnoughSpaceError,
   createUnhandledError,
 } from './processingErrors.js';
+import { CID } from 'multiformats';
 
 interface ProcessS3DataToIpfsParams {
   files: any[];
@@ -638,7 +639,7 @@ export async function handleCleanupOnMidProcessingError({
       size: e.size || 0,
       nodeId: node.id,
       userId: user.id,
-      directory: await isDir(e.cid),
+      directory: await isDir(CID.parse(e.cid)),
     };
   });
   const prunedEntries = await prisma.cidPruneList.createMany({ data: await Promise.all(formattedPruneList) });
